requirements.txt
streamlit
numpy==1.23.2
pandas-datareader==0.10.0
pandas-ta==0.3.14b0
streamlit==1.12.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
yfinance==0.1.74
cufflinks==0.17.3